+++ src/main/java/GUI/Table/DoctorsOfPolyclinicsTable.java
@@ -8,18 +8,28 @@ import GUI.Row.Mode;
 import java.util.Arrays;
 import java.util.Vector;
 
-public class DoctorsOfPolyclinicsTable extends TableView{
-    String [] nameColumns = {"ID", "Больница", "ФИО врача"};
+public class DoctorsOfPolyclinicsTable extends TableView {
+    String[] nameColumns = {"ID", "Больница", "ФИО врача"};
+
     public DoctorsOfPolyclinicsTable(String name, String userID, Role role) {
         super(name, userID, role);
+        if (role == Role.PATIENT) addButton.setVisible(false);
         updateTable();
     }
 
     @Override
     public void updateTable() {
-        Vector values = ConnectionManager.select("SELECT id, p.name, d.surname || ' ' || d.name || ' ' || d.patronymic " +
-                "FROM doctors_of_polyclinics JOIN polyclinics p USING (polyclinic_id) " +
-                "JOIN doctors d USING (doctor_id) ORDER BY id", 3);
+        Vector values;
+        if (role == Role.POlYCLINIC_REGISTRY) {
+            values = ConnectionManager.select("SELECT id, p.name, d.surname || ' ' || d.name || ' ' || d.patronymic " +
+                    "FROM doctors_of_polyclinics JOIN polyclinics p USING (polyclinic_id) " +
+                    "JOIN doctors d USING (doctor_id) JOIN users_polyclinics USING (polyclinic_id) WHERE " +
+                    "(user_id = " + userID + ") ORDER BY id", 3);
+        } else {
+            values = ConnectionManager.select("SELECT id, p.name, d.surname || ' ' || d.name || ' ' || d.patronymic " +
+                    "FROM doctors_of_polyclinics JOIN polyclinics p USING (polyclinic_id) " +
+                    "JOIN doctors d USING (doctor_id) ORDER BY id", 3);
+        }
         Vector header = new Vector(Arrays.asList(nameColumns));
         dtm.setDataVector(values, header);
         table.setModel(dtm);
@@ -37,6 +47,7 @@ public class DoctorsOfPolyclinicsTable extends TableView{
 
     @Override
     public void deleteRow(Integer id) {
-        ConnectionManager.delete("DELETE FROM doctors_of_polyclinics WHERE id = " + id);
+        if (role == Role.ADMIN || role == Role.POlYCLINIC_REGISTRY)
+            ConnectionManager.delete("DELETE FROM doctors_of_polyclinics WHERE id = " + id);
     }
 }
