+++ src/main/java/GUI/Table/PatientsTable.java
@@ -12,13 +12,20 @@ public class PatientsTable extends TableView{
     String [] nameColumns = {"ID", "Фамилия", "Имя", "Отчество", "Дата рождения", "Телефон"};
     public PatientsTable(String name, String userID, Role role) {
         super(name, userID, role);
+        if (role == Role.PATIENT) addButton.setVisible(false);
         updateTable();
     }
 
     @Override
     public void updateTable() {
-        Vector values = ConnectionManager.select("SELECT patient_id, surname, name, patronymic, birthday, phone_number \n" +
-                "FROM patients", 6);
+        Vector values;
+        if (role == Role.PATIENT) {
+            values = ConnectionManager.select("SELECT patient_id, surname, name, patronymic, birthday, phone_number " +
+                    "FROM patients JOIN users_patients USING (patient_id) WHERE (user_id = " + userID + ")", 6);
+        } else {
+            values = ConnectionManager.select("SELECT patient_id, surname, name, patronymic, birthday, phone_number " +
+                    "FROM patients", 6);
+        }
         Vector header = new Vector(Arrays.asList(nameColumns));
         dtm.setDataVector(values, header);
         table.setModel(dtm);
@@ -26,15 +33,18 @@ public class PatientsTable extends TableView{
 
     @Override
     public void createNewRow() {
-        PatientRowView add = new PatientRowView("Добавление врача", Mode.CREATE);
+        PatientRowView add = new PatientRowView("Добавление пациента", Mode.CREATE);
     }
 
     @Override
     public void editRow(Integer id) {
+        PatientRowView edit = new PatientRowView("Изменение пациента", Mode.EDIT);
+        edit.setSelectedRow(id);
+        edit.fillFields();
     }
 
     @Override
     public void deleteRow(Integer id) {
-        ConnectionManager.delete("DELETE FROM patients WHERE patient_id = " + id);
+        if (role!=Role.PATIENT) ConnectionManager.delete("DELETE FROM patients WHERE patient_id = " + id);
     }
 }
